Move Verification process to TrianlugarSolver.verify

+++ dataStructure/include/SparseMatrix.h
@@ -101,7 +101,7 @@ namespace DataStructure {
             return this->Lx;
         }
 
-        long getSize() {
+        int getSize() {
             return this->M;
         }
